backend-services/services/brain-orchestrator/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-cloud-pubsub==2.18.4
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0
google-cloud-bigtable==2.21.0
google-cloud-secretmanager==2.16.3
psycopg2-binary==2.9.7
redis==5.0.1
pytest==7.4.0  
pytest-mock==3.11.1 
pytest==7.4.0  
pytest-mock==3.11.1 
